ahora hay un int

+++ src/main/java/App.java
@@ -3,14 +3,14 @@ import java.util.Scanner;
 public class App {
     public static void main (String[] args) {
         Scanner entrada=new Scanner(System.in);//recoje información que pasa el usuario por el teclado
-        double precioTotal=0;
-        double precioArticulo;
+        int precioTotal=0;
+        int precioArticulo;
         String hayMasArticulos;
         do{ //alinear las llaves
             System.out.print("Ingrese el precio del artículo: ");
             //precioArticulo = input.nextint();
             // input utilizado no existe
-            precioArticulo = entrada.nextDouble();
+            precioArticulo = entrada.nextInt();
             while (precioArticulo < 0) {
                 System.out.print("Precio inválido. Ingrese un precio positivo: ");
                 precioArticulo = entrada.nextInt();
@@ -21,8 +21,8 @@ public class App {
         } while(hayMasArticulos.equals("S"));
         System.out.println("El total a pagar es: " + precioTotal + " €");
         System.out.print("Ingrese el dinero entregado por el cliente: ");
-        double dineroADar = entrada.nextDouble();
-        double cambio = dineroADar - precioTotal;
+        int dineroADar = entrada.nextDouble();
+        int cambio = dineroADar - precioTotal;
         System.out.println("El total a pagar son " + precioTotal + " €, el cliente entrega "
                 + dineroADar + " €, por lo que el cambio es " + cambio + " €.");
     }
